Use int64_t instead of long long

+++ AppCenterAnalytics/AppCenterAnalytics/Internals/Model/Properties/MSLongTypedProperty.h
@@ -6,6 +6,6 @@
 /**
  * Long property value (64-bit signed integer).
  */
-@property(nonatomic) long long value;
+@property(nonatomic) int64_t value;
 
 @end
